Demonstrate a bimap in ZDDs.
This map is a bimap that holds one copy of each key-value pair, yet lookup is
still efficient.

The map is also a multimap, each key holds a set of values. There's an API for
efficiently finding keys that hold the same value.

There's also an API which allows one key to "steal" the values from a set of
other keys.